ref(in-call): encapsulate meeting toolbar as separate component
The intent being to be able to choose which of various
types of meeting toolbars to display, as some meeting
providers may have different functionality. For example,
zoom would have a simpler toolbar.
