+++ spot-client/src/spot-remote/ui/components/meeting-toolbar/MeetingToolbar.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import JitsiMeetingToolbar from './JitsiMeetingToolbar';
+
+/**
+ * A component for showing in-meeting controls for a Spot-Remote to control
+ * the meeting on the Spot-TV.
+ *
+ * @returns {ReactElement}
+ */
+export default function MeetingToolbar() {
+    return <JitsiMeetingToolbar />;
+}
