+++ spot-client/src/spot-remote/ui/components/meeting-toolbar/buttons/AudioMuteButton.test.js
