+++ spot-client/src/spot-remote/ui/components/meeting-toolbar/buttons/VideoMuteButton.test.js
