+++ spot-client/src/spot-remote/ui/components/meeting-toolbar/index.js
@@ -1,0 +1,1 @@
+export { default as MeetingToolbar } from './MeetingToolbar';
