+++ spot-client/src/spot-remote/ui/components/nav/index.js
@@ -1,4 +1,3 @@
-export * from './buttons';
 export { default as NavButton } from './nav-button';
 export { default as NavContainer } from './nav-container';
 export { default as NavItem } from './NavItem';
